final-pjt-back/requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.9
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
